% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint.R
\name{Add_Constraint}
\alias{Add_Constraint}
\title{Create a Linear Constraint}
\usage{
Add_Constraint(Parcels, Coefficients, RHS,
                      Silence=FALSE, Env= .GlobalEnv)
}
\arguments{
\item{Parcels}{A vector of strings: the name of the parcels to appear in the constraint.}

\item{Coefficients}{A vector of doubles: the respective coefficients of the \code{Parcels}.}

\item{RHS}{A double: defining the right-hand-side value of the constraint.}

\item{Silence}{A binary parameter: if \code{FALSE}, the function will report a summary message. If \code{TRUE}, the function will be silent.  (Default: FALSE)}

\item{Env}{the environment where the package should create or access variables.
By default the package works in the R's Global environment. (Default: .GlobalEnv)}
}
\value{
NULL (creates some variables in the environment but returns nothing)
}
\description{
\code{Add_Constraint} creates a Constraint in the optimization model of the form:\cr
\eqn{Coefficients . Parcels \le RHS},\cr
where \code{Coefficients} is a vector of doubles, \code{Parcels} is the vector of parcel names, and  \code{RHS} is a double representing the right-hand-side value of the constraint.
}
\details{
When used for the first time, \code{Add_Constraint} will create a list named \code{Constraints}.

Each member of the list \code{Constraints} represents a linear constraint.
Each constraint is stored as a list with three members: \code{Parcels} storing the names of the parcels in the constraint,
\code{Coefficients} storing the respective coefficients of the parcels, and \code{RHS} storing the constraints right-hand-side value.

In other words, the information of the \code{i}'th constraint in \code{Constraints} can be accessed by \cr
\code{Constraints[[i]][["Parcels"]]}, \cr
\code{Constraints[[i]][["Coefficients"]]}, \cr
and \code{Constraints[[i]][["RHS"]]}.

This function enables adding any types of linear constraints to the optimization model.

\strong{Removing a constraint:} Constraint \code{i} can be removed using the following code:

\code{Constraints[[i]]<-NULL}

\strong{Remove all Constraints:} All of the Constraints stored in \code{Constraints} can be removed using the following code:

\code{remove(Constraints)}
}
\examples{
\dontrun{

## The constraint  {Parcel_1 + Parcel_3 + Parcel_5 >= 2}
is the same as  {-Parcel_1 - Parcel_3 - Parcel_5 <= -2},
which can be created by using the following code: ##

Add_Constraint(Parcels= c("Parcel_1","Parcel_3","Parcel_5"),
               Coefficients= c(-1,-1,-1),
               RHS=-2,
               Silence=FALSE)

}









}
\seealso{
Other Constraints: 
\code{\link{Budget_Add_Constraint}()},
\code{\link{Connectivity_Add_Constraint}()},
\code{\link{Disconnectivity_Add_Constraint}()}
}
\concept{Constraints}
