% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Budget.R
\name{Budget_Add_Constraint}
\alias{Budget_Add_Constraint}
\title{Create a Budget Constraint}
\usage{
Budget_Add_Constraint(Parcels, Coefficients, Limit, Silence=FALSE, Env= .GlobalEnv)
}
\arguments{
\item{Parcels}{A vector of strings: the name of the parcels appearing in the constraint.}

\item{Coefficients}{A vector of doubles: the respective coefficients of the parcels in \code{Parcels}.}

\item{Limit}{A double: defining the budget limit (right-hand-side)}

\item{Silence}{A binary parameter: if \code{FALSE}, the function will report a summary message. If \code{TRUE}, the function will be silent.  (Default: FALSE)}

\item{Env}{the environment where the package should create or access variables.
By default the package works in the R's Global environment. (Default: .GlobalEnv)}
}
\value{
NULL (creates some variables in the environment but returns nothing)
}
\description{
\code{Budget_Add_Constraint} creates a Budget Constraint in the optimization model of the form:\cr
\eqn{Coefficients . Parcels \le Limit}\cr
where \code{Coefficients} is a vector of doubles, \code{Parcels} is the vector of parcel names, and  \code{Limit} is a double representing the budget limit.
}
\details{
When used for the first time, \code{Budget_Add_Constraint} or \link{Budget_Import_Constraint} will create a list named \code{Budget} in the environment.
Whenever a new budget constraint is added, a new member will be added to the list \code{Budget}.

Each member of the list \code{Budget} represents a budget constraint.
Each budget constraint is stored as a list with three members: \code{Parcels} storing the names of the parcels in the constraint,
\code{Coefficients} storing the respective coefficients of the parcels, and \code{Limit} storing the budget limit (right-hand-side).\cr

In other words, the information of the \code{i}'th Budget constraint in \code{Budget} can be accessed by:\cr
\code{Budget[[i]][["Parcels"]]},\cr
\code{Budget[[i]][["Coefficients"]]},\cr
and \code{Budget[[i]][["Limit"]]}.

\strong{Removing a budget:} Budget \code{i} can be removed using the following code:

\code{Budget[[i]]<-NULL}

\strong{Removing all budget constraints:} All budget constraints can be removed using the following code:

\code{remove(Budget)}
}
\examples{
\dontrun{

## To create a budget constraint of the form
5*Parcel_1 + 10*Parcel_3 + 15*Parcel_4 <= 15 ##

Budget_Add_Constraint(Parcels= c("Parcel_1","Parcel_3","Parcel_4"),
                      Coefficients= c(5,10,15), Limit=15)

}


}
\seealso{
Other Budget Constraints: 
\code{\link{Budget_Import_Constraint}()}

Other Constraints: 
\code{\link{Add_Constraint}()},
\code{\link{Connectivity_Add_Constraint}()},
\code{\link{Disconnectivity_Add_Constraint}()}
}
\concept{Budget Constraints}
\concept{Constraints}
