% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disconnectivity.R
\name{Disconnectivity_Add_Constraint}
\alias{Disconnectivity_Add_Constraint}
\title{Create a Disconnectivity Constraint}
\usage{
Disconnectivity_Add_Constraint(Target, Condition, Repetition = TRUE,
                                      Silence=FALSE, Env= .GlobalEnv)
}
\arguments{
\item{Target}{A string: the name of a parcel which can only be protected (invested) if none of the \code{Condition} parcels are protected (invested).}

\item{Condition}{A vector of strings: the names of the parcels that if at least one of them are protected (invested), the \code{Target} cannot be protected (invested).}

\item{Repetition}{A binary parameter: If \code{TRUE}, the \code{Condition} parcels will create a new constraint for \code{Target}.
If \code{FALSE}, the \code{Condition} parcels will be added to the \bold{first} existing disconnectivity constraint of \code{Target}, i.e. \code{Disconnectivity[["Target"]][[1]]}. (Default: TRUE)}

\item{Silence}{A binary parameter: if \code{FALSE}, the function will report a summary message. If \code{TRUE}, the function will be silent.  (Default: FALSE)}

\item{Env}{the environment where the package should create or access variables.
By default the package works in the R's Global environment. (Default: .GlobalEnv)}
}
\value{
NULL (creates some variables in the environment but returns nothing)
}
\description{
\code{Disconnectivity_Add_Constraint} creates a Disconnectivity Constraint in the optimization model of the form:\cr
\eqn{\sum (Condition)  + C * Target \le C},\cr
where \code{C} is the number of parcels in \code{Condition}. This constraint means that the parcel \code{Target} can be protected (invested)
\bold{only if} none of the parcels in \code{Condition} are protected (invested).
}
\details{
When used for the first time, \code{Disconnectivity_Add_Constraint} or\cr
\link{Disconnectivity_Import_Constraint} will create a list named \code{Disconnectivity}.
When, for the first time, a disconnectivity constraint is added for a \code{Target} parcel, for instance named "Parcel_1",
a new member named "Parcel_1" will be added to the \code{Disconnectivity} list, i.e. \code{Disconnectivity[["Parcel_1"]]}.

Then, if \code{Repetition=FALSE}, whenever \code{Disconnectivity_Add_Constraint} or\cr
\link{Disconnectivity_Import_Constraint} are used for "Parcel_1", the \code{Condition} parcels will be added to "Parcel_1"'s
first disconnectivity constraint, i.e. \code{Disconnectivity[["Parcel_1"]][[1]]}.
Otherwise, if \code{Repetition=TRUE}, the \code{Condition} parcels will create a new constraint for "Parcel_1", i.e.\cr
\code{Disconnectivity[["Parcel_1"]][[i+1]]}, where \code{i} is the number of "Parcel_1"'s existing disconnectivity constraints.

\strong{Removing a specific disconnectivity constraint of a \code{Target}:} Disconnectivity constraint \code{i} of a parcel named "Parcel_j"
can be removed using the following code:

\code{Disconnectivity[["Parcel_j"]][[i]]<-NULL}

\strong{Removing all disconnectivity constraints of a \code{Target}:} Disconnectivity constraints of a parcel named "Parcel_j"
can be removed using the following code:

\code{Disconnectivity[["Parcel_j"]]<-NULL}

\strong{Remove all Disconnectivity Constraints:} All of the Disconnectivity constraints can be removed using the following code:

\code{remove(Disconnectivity)}
}
\examples{
\dontrun{

Disconnectivity_Add_Constraint(Target= "Parcel_1", Condition= "Parcel_4", Repetition=TRUE)

## will create Disconnectivity[["Parcel_1"]][[1]] and
add the following constraint to the optimization model
Parcel_4 + Parcel_1 <= 1 . Now, if the code ##

Disconnectivity_Add_Constraint(Target= "Parcel_1", Condition= c("Parcel_2","Parcel_3"),
                               Repetition=TRUE)

## is run, Disconnectivity[["Parcel_1"]][[2]] will be created,
and the following constraint will be added to the optimization model
Parcel_2 + Parcel_3 + 2*Parcel_1 <= 2 . However, if the code ##

Disconnectivity_Add_Constraint(Target= "Parcel_1", Condition= c("Parcel_2","Parcel_3"),
                               Repetition=FALSE)

## is run, only Disconnectivity[["Parcel_1"]][[1]] will be updated,
and its existing constraint will be updated in the optimization model as
Parcel_4 + Parcel_2 + Parcel_3 + 3*Parcel_1 <= 3 . ##

}



}
\seealso{
Other Disconnectivity Constraints: 
\code{\link{Disconnectivity_Import_Constraint}()}

Other Constraints: 
\code{\link{Add_Constraint}()},
\code{\link{Budget_Add_Constraint}()},
\code{\link{Connectivity_Add_Constraint}()}
}
\concept{Constraints}
\concept{Disconnectivity Constraints}
