% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_data.R
\name{Import_data}
\alias{Import_data}
\title{Import problem's data from Excel file}
\usage{
Import_data(Address, First_Objective_Sense= "Min",
                   Second_Objective_Sense="Min", Silence= FALSE, Env= .GlobalEnv)
}
\arguments{
\item{Address}{A string: the location and name of the excel file.}

\item{First_Objective_Sense}{A string: "Min" means the objective is to be minimized and "Max" means the objective is to be maximized. (Default: "Min")}

\item{Second_Objective_Sense}{A string: "Min" means the objective is to be minimized and "Max" means the objective is to be maximized. (Default: "Min")}

\item{Silence}{A binary parameter: if \code{FALSE}, the function will report a summary message. If \code{TRUE}, the function will be silent.  (Default: FALSE)}

\item{Env}{the environment where the package should create or access variables.
By default the package works in the R's Global environment. (Default: .GlobalEnv)}
}
\value{
NULL (creates some variables in the environment but returns nothing)
}
\description{
\code{Import_data} imports the problem's data from an excel file.
The excel file must include a sheet named "Parcels" providing names of all parcels in the problem and their status.\cr
If any of the sheets named "First_Objective", "Second_Objective", "Sigma", and "Correlation" exist in the excel file, the function will automatically import the data from them.\cr
\strong{Note:} The sheets in the excel file must follow the structures explained in the functions in \verb{See Also} section.
}
\details{
Based on the sheets available, \code{Import_data} will automatically call different functions.

If "Parcels" exists, the function \link{Parcels_Import} will be called. It will create
variables \code{Parcels} representing the name of parcels and the order they are used,
and \code{Status} showing which parcel is or is not protected (owned).\cr
(For more information refer to function \link{Parcels_Import})

If "First_Objective" exists, the function \link{First_Objective_Import} will be called, which will create
a list named \code{First_Objective}. \code{First_Objective[["Coefficients"]]} stores the the coefficient of parcels
and \code{First_Objective[["Sense"]]} stores the parameter \code{First_Objective_Sense}
showing if the objective is to be minimized or maximized.\cr
(For more information refer to function \link{First_Objective_Import})

If "Second_Objective" exists, the function \link{Second_Objective_Import} will be called, which will create
a list named \code{Second_Objective}. \code{Second_Objective[["Coefficients"]]} stores the the coefficient of parcels
and \code{Second_Objective[["Sense"]]} stores parameter \code{Second_Objective_Sense}
showing if the objective is to be minimized or maximized.\cr
(For more information refer to function \link{Second_Objective_Import})

If both "Sigma" and "Correlation" exist, the function \link{Risk_Objective_Import} will be called.
It will creat a list named \code{Risk_Objective} will be created,
where \code{Risk_Objective[["Sigma"]]} stores the the sigmas of parcels,
\code{Risk_Objective[["Correlation"]]} stores the the Correlation of parcels.\cr
(For more information refer to function \link{Risk_Objective_Import})
}
\examples{
\dontrun{

Import_data(Address= "C:\\\\Folder\\\\example.xlsx",
            First_Objective_Sense= "Max", Second_Objective_Sense="Max")
}



}
\seealso{
Other Import Functions: 
\code{\link{First_Objective_Import}()},
\code{\link{Parcels_Import}()},
\code{\link{Risk_Objective_Import}()},
\code{\link{Second_Objective_Import}()}
}
\concept{Import Functions}
