% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{Optimize_Second_Objective}
\alias{Optimize_Second_Objective}
\title{Load and Optimize the second Objective Function}
\usage{
Optimize_Second_Objective(Time_limit=1e7, Solver="SCIP",
                                 Silence= FALSE, Env= .GlobalEnv)
}
\arguments{
\item{Time_limit}{A double: the total time limit in seconds}

\item{Solver}{A string: defining the solver to be used to solve the problem. (Default: "SCIP")}

\item{Silence}{A binary parameter: if \code{FALSE}, the function will report a summary message. If \code{TRUE}, the function will be silent.  (Default: FALSE)}

\item{Env}{the environment where the package should create or access variables.
By default the package works in the R's Global environment. (Default: .GlobalEnv)}
}
\value{
A list with 5 members:\cr
\verb{[["Solution_time"]]}: a numeric value; defining the time spent in the solver in seconds.\cr
\verb{[["Status"]]}: a character; defining the status of the solution.\cr
\verb{[["Gap"]]}: a numeric value; defining the relative optimality gap of the solution.\cr
\verb{[["Result"]]}: a vector of binaries; its \code{i}'th element defines if the parcel named \code{Parcels[i]} is protected or not.\cr
\verb{[["Objective_Value"]]}: a numveric value; the returned value of Second objective.
}
\description{
\code{Optimize_Second_Objective} loads the optimization problem and optimizes \code{Second_Objective} in Julia environment.\cr
}
\details{
In the list that \code{Optimize_Second_Objective} returns, \verb{[["Status"]]} defines the status of the returned solution.\cr
If \verb{[["Status"]]="OPTIMAL"}, the problem was solved to optimality and the parcels' optimal status are stored in \verb{[["Result"]]}.
Also, the optimal value will be stored in \verb{[["Objective_Value"]]}.\cr
If \verb{[["Status"]]="TIME_LIMIT"}, the solver was terminated because the time limit was reached.
If any feasible solution was reported by the solver, it would be stored in \verb{[["Result"]]}. Otherwise, \verb{[["Result"]]} and \verb{[["Objective_Value"]]} would be
both equal to "N/A".\cr
If \verb{[["Status"]]="INFEASIBLE"}, the problem is infeasible. So, no solution would be reported.\cr

In any other case, the \verb{[["Status"]]} will be the solution status that the solver has reported.
}
\examples{
\dontrun{

Optimization_Result = Optimize_Second_Objective(Time_limit=1e7, Solver="SCIP")
}

}
\seealso{
Other Optimizer Functions: 
\code{\link{Load_Problem}()},
\code{\link{Optimize_First_Objective}()},
\code{\link{Optimize_NBP}()},
\code{\link{Optimize_WSO}()}

Other Second Objective: 
\code{\link{Second_Objective_Create}()},
\code{\link{Second_Objective_Import}()}
}
\concept{Optimizer Functions}
\concept{Second Objective}
