% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_WSO.R
\name{Optimize_WSO}
\alias{Optimize_WSO}
\title{Load and Optimize Weighted Sum Optimization (WSO)}
\usage{
Optimize_WSO(First_Objective=0, Second_Objective=0, Risk_Objective=0,
                    Time_limit=1e7, Solver="SCIP", Silence= FALSE, Env= .GlobalEnv)
}
\arguments{
\item{First_Objective}{A float parameter: defining the weight of the first objective function (\code{First_Objective}) in the optimization.
If \bold{0}, it will \strong{not} be included in optimization.}

\item{Second_Objective}{A float parameter: defining the weight of the second objective function (\code{Second_Objective}) in the optimization.
If \bold{0}, it will \strong{not} be included in optimization.}

\item{Risk_Objective}{A float parameter: defining the weight of the Risk objective function (\code{Risk_Objective}) in the optimization.
If \bold{0}, it will \strong{not} be included in optimization.}

\item{Time_limit}{A double: the total time limit in seconds}

\item{Solver}{A string: defining the solver to be used to solve the problem. (Default: "SCIP")}

\item{Silence}{A binary parameter: if \code{FALSE}, the function will report a summary message. If \code{TRUE}, the function will be silent.  (Default: FALSE)}

\item{Env}{the environment where the package should create or access variables.
By default the package works in the R's Global environment. (Default: .GlobalEnv)}
}
\value{
A list with 8 members:\cr
\verb{[["Solution_time"]]}: a numeric value; defining the time spent in the solver in seconds.\cr
\verb{[["Status"]]}: a character; defining the status of the solution.\cr
\verb{[["Gap"]]}: a numeric value; defining the relative optimality gap of the solution.\cr
\verb{[["Result"]]}: a vector of binaries; its \code{i}'th element defines if the parcel named \code{Parcels[i]} is protected or not.\cr
\verb{[["Firs_Objective"]]}: a numveric value; (if applicable) returns the WSO value of first objective.\cr
\verb{[["Second_Objective"]]}: a numveric value; (if applicable) returns the WSO value of second objective.\cr
\verb{[["Risk_Objective"]]}: a numveric value; (if applicable) returns the WSO value of risk objective.\cr
}
\description{
\code{Optimize_WSO} creates and optimizes the Weighted Sum Optimization (WSO) in Julia environment.
It requires two objective functions and their respective positive weights.\cr
(For detailed information refer to the paper)
}
\details{
In the list that \code{Optimize_WSO} returns, \verb{[["Status"]]} defines the status of the returned WSO solution.\cr
If \verb{[["Status"]]="OPTIMAL"}, the WSO was solved to optimality and the Parcels' optimal status are stored in \verb{[["Result"]]}.
Also, the WSO optimal values of objectives will be stored in either \verb{[["First_Objective"]]}, \verb{[["Second_Objective"]]}, or \verb{[["Risk_Objective"]]}
, based on the input objectives.\cr
If \verb{[["Status"]]="TIME_LIMIT"}, the solver was terminated because the time limit was reached.
If any feasible solution was reported by the solver, it would be stored in \verb{[["Result"]]}. Otherwise, \verb{[["Result"]]} and objective values would be
all equal to "N/A".\cr
If \verb{[["Status"]]="INFEASIBLE"}, the problem is infeasible. So, no solution would be reported.\cr

In any other case, the \verb{[["Status"]]} will be the solution status that the solver has reported.

\strong{Note 1:} The weights should be non negative, and two objectives should have positive weights.\cr
}
\examples{
\dontrun{

## to find WSO solution for First Objective and Second Objective with weights equal to 1 and 2: ##

Optimize_WSO(First_Objective=1, Second_Objective=2, Time_limit=1e7, Solver="SCIP", Silence= FALSE)

## to find WSO solution for First Objective and Risk Objective with weights equal to 1 and 2: ##

Optimize_WSO(First_Objective=1, Risk_Objective=2, Time_limit=1e7, Solver="SCIP", Silence= FALSE)

## to find WSO solution for Second Objective and Risk Objective with weights equal to 1 and 2: ##

Optimize_WSO(Second_Objective=1, Risk_Objective=2, Time_limit=1e7, Solver="SCIP", Silence= FALSE)
}


}
\seealso{
Other Optimizer Functions: 
\code{\link{Load_Problem}()},
\code{\link{Optimize_First_Objective}()},
\code{\link{Optimize_NBP}()},
\code{\link{Optimize_Second_Objective}()}
}
\concept{Optimizer Functions}
