% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parcels_status.R
\name{Parcels_Create}
\alias{Parcels_Create}
\title{Create Parcels and their Status}
\usage{
Parcels_Create(Parcels, Status, Silence= FALSE, Env= .GlobalEnv)
}
\arguments{
\item{Parcels}{A vector of strings: the name of the parcels appearing in the problem}

\item{Status}{A vector of binary: the respective status of the parcels in \code{Parcels}.
Value \bold{1} means the parcel is protected (owned), and value \bold{0} means the
parcel is unprotected (unowned).}

\item{Silence}{A binary parameter: if \code{FALSE}, the function will report a
summary message. If \code{TRUE}, the function will be silent.  (Default: FALSE)}

\item{Env}{the environment where the package should create or access variables.
By default the package works in the R's Global environment. (Default: .GlobalEnv)}
}
\value{
NULL (creates some variables in the environment but returns nothing)
}
\description{
\code{Parcels_Create} manually creates the problem's data. It
will create variables named \code{Parcels} storing the name of parcels and the order they are used, and
\code{Status} showing whether parcels are protected or not.\cr
\strong{Note:} These variables are essential to create the optimization
problem. So, if not created, calling other functions will result in an error.
}
\details{
\code{Parcels} should include the name of all parcels appearing in the problem,
and \code{Status} should include all the parcels' status, where value
\bold{1} means the parcel is protected (owned), and value \bold{0} means the
parcel is unprotected (unowned). These values are essential in creating
optimization variables.\cr

In the optimization model, protected (owned) and unprotected (unowned) parcels
will be represented by the binary decision variables \code{(1-y)} and
\code{x}. In the final solution after optimization, \code{(1-y)=1} means that
the parcel which was initially protected (owned) remains protected, and
\code{(1-y)=0} means that the parcel is divested (sold).\cr Similarly, \code{x=1} means
that the parcel that was initially unprotected (unowned) will become protected (invested), and
\code{x=0} means that the parcel will remain unprotected.\cr (For more
information refer to the Paper)
}
\examples{
\dontrun{

Parcels_Create(Parcels=c("Parcel_1","Parcel_2","Parcel_3","Parcel_4","Parcel_5"),
                  Status= c(0,1,0,1,1))
}


}
\seealso{
Other Parcels Status: 
\code{\link{Parcels_Import}()}
}
\concept{Parcels Status}
