% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_Objective.R
\name{Risk_Objective_Create}
\alias{Risk_Objective_Create}
\title{Create Risk Objective}
\usage{
Risk_Objective_Create(Parcels, Sigma, Correlation,
                             Silence=FALSE, Env= .GlobalEnv)
}
\arguments{
\item{Parcels}{A vector of strings: the name of the parcels appearing in the objective}

\item{Sigma}{A vector of doubles: the respective sigmas of the parcels in \code{Parcels}.}

\item{Correlation}{A matrix of doubles: a symmetric matrix with diogonal elements equal to one.
The element in \code{i}'th row and \code{j}'th column is the correlation between the \code{i} and \code{j} member of \code{Parcels}.}

\item{Silence}{A binary parameter: if \code{FALSE}, the function will report a summary message. If \code{TRUE}, the function will be silent.  (Default: FALSE)}

\item{Env}{the environment where the package should create or access variables.
By default the package works in the R's Global environment. (Default: .GlobalEnv)}
}
\value{
NULL (creates some variables in the environment but returns nothing)
}
\description{
\code{Risk_Objective_Create} creates the problem's Risk objective function.
It creates a list named \code{Risk_Objective}, where \code{Risk_Objective[["Sigma"]]} stores the the sigmas of parcels, and\cr
\code{Risk_Objective[["Correlation"]]} stores the the Correlation of parcels.\cr
This function is not sensitive to the order the parcels are stored in the sheet and will automatically store parcels sorted with the same order they were stored in \code{Parcels}.
}
\examples{
\dontrun{

Risk_Objective_Create(Parcels = c("Parcel_1","Parcel_2","Parcel_3"),
                      Sigma = c(1,2,3),
                     Correlation = cbind(c(1,0.5,-0.7),c(0.5,1,0.2),c(-0.7,0.2,1)))

## Correlation matrix in this example is
 | +1.0 | +0.5 | -0.7 |
 | +0.5 | +1.0 | +0.2 |
 | -0.7 | +0.2 | +1.0 | ##
}


}
\seealso{
Other Risk Objective: 
\code{\link{Risk_Objective_Import}()}

Other Objective Creators: 
\code{\link{First_Objective_Create}()},
\code{\link{Second_Objective_Create}()}
}
\concept{Objective Creators}
\concept{Risk Objective}
