% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install.R
\name{Setup}
\alias{Setup}
\title{One-time installation function}
\usage{
Setup()
}
\description{
This function should be called only once when installing the package for the first time. This function should be called after when Julia 1.3 and SCIP 6.0.2 are installed and their global variables are defined.
}
\details{
This function will initialize the installation of required packages and libraries in Julia. The packages are "JuMP.jl" and "SCIP.jl".
The installation may take several minutes to complete, and when completed, the message "Setup Completed!" will appear.
}
